parametrize attaching when launching a process

+++ src/process.cpp
@@ -17,7 +17,7 @@ namespace {
 }
 
 std::unique_ptr<vdb::process> vdb::process::launch(
-		std::filesystem::path path) {
+		std::filesystem::path path, bool debug) {
 	// gotta do this before forking,
 	// otherwise the child and parent have their own (different) pipes
 	pipe channel(/*close_on_exec=*/true);
@@ -30,7 +30,7 @@ std::unique_ptr<vdb::process> vdb::process::launch(
 	if (pid == 0) {
 		// the child just writes and doesn't read
 		channel.close_read();
-		if (ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0) {
+		if (debug and ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0) {
 			exit_with_perror(channel, "Tracing failed");
 		}
 		if (execlp(path.c_str(), path.c_str(), nullptr) < 0) {
@@ -49,8 +49,10 @@ std::unique_ptr<vdb::process> vdb::process::launch(
 		error::send(std::string(chars, chars + data.size()));
 	}
 
-	std::unique_ptr<process> proc (new process(pid, /*terminate_on_end=*/true));
-	proc->wait_on_signal();
+	std::unique_ptr<process> proc (new process(pid, /*terminate_on_end=*/true, debug));
+	if (debug) {
+		proc->wait_on_signal();
+	}
 
 	return proc;
 }
@@ -65,7 +67,7 @@ std::unique_ptr<vdb::process> vdb::process::attach(
 		error::send_errno("Could not attach");
 	}
 
-	std::unique_ptr<process> proc (new process(pid, /*terminate_on_end=*/false));
+	std::unique_ptr<process> proc (new process(pid, /*terminate_on_end=*/false, /*attached=*/true));
 	proc->wait_on_signal();
 
 	return proc;
@@ -74,16 +76,18 @@ std::unique_ptr<vdb::process> vdb::process::attach(
 vdb::process::~process() {
 	if (pid_ != 0) {
 		int status;
-		if (state_ == process_state::running) {
-			kill(pid_, SIGSTOP);
-			waitpid(pid_, &status, 0);
-		}
-		ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
-		kill(pid_, SIGCONT);
+		if (is_attached_) {
+			if (state_ == process_state::running) {
+				kill(pid_, SIGSTOP);
+				waitpid(pid_, &status, 0);
+			}
+			ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
+			kill(pid_, SIGCONT);
 
-		if (terminate_on_end_) {
-			kill(pid_, SIGKILL);
-			waitpid(pid_, &status, 0);
+			if (terminate_on_end_) {
+				kill(pid_, SIGKILL);
+				waitpid(pid_, &status, 0);
+			}
 		}
 	}
 }
